Add confirmation alert when deleting a row from the list.

+++ frontend/src/components/DataBlockAdmin.js
@@ -1,10 +1,18 @@
 import React, { useEffect, useState } from "react";
 import { IconButton, Button } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogContentText from "@mui/material/DialogContentText";
+import DialogTitle from "@mui/material/DialogTitle";
+import AlertDialog from "./AlertDialog";
+
 const axios = require("axios").default;
 
 const DataBlockAdmin = (props) => {
   const [visible, setVisible] = useState(true);
+  const [display, setDisplay] = useState(false);
 
   const deleteItem = async () => {
     console.log(props.id);
@@ -14,6 +22,14 @@ const DataBlockAdmin = (props) => {
     console.log(response);
     setVisible(false);
   };
+  const openConfirmation = () => {
+    setDisplay(true);
+  };
+
+  const stopDisplay = () => {
+    setDisplay(false);
+  };
+
   return (
     <div>
       {visible ? (
@@ -24,15 +40,23 @@ const DataBlockAdmin = (props) => {
           <Button
             variant="outlined"
             color="error"
-            onClick={deleteItem}
+            onClick={openConfirmation}
             startIcon={<DeleteIcon />}
           >
             Delete
           </Button>
+          <AlertDialog
+            confirmDelete={openConfirmation}
+            stopDisplay={stopDisplay}
+            deleteItem={deleteItem}
+            display={display}
+            finnish={props.finnish}
+            english={props.english}
+            category={props.category}
+          />
         </div>
       ) : null}
     </div>
   );
 };
-
 export default DataBlockAdmin;
